QuantumOrderEffectModels.QOEM — Type

QOEM{T<:Real,V<:AbstractVector{T}} <: AbstractQOEM

A model object for a quantum model for medical diagnosis.

Basis vectors

A person's cognitive state Ψ is represented in a 4D belief space where basis vectors correspond to hypotheses and evidence states:

disease present and positive evidence for disease

disease present and negative evidence for disease

disease absent and positive evidence for disease

disease absent and negative evidence for disease

Fields

Ψ::V: initial state vector (superposition)

γₕ::T: rotation for positive evidence from medical history

γₗ::T: rotation for negative evidence from laboratory test

σ::T: the standard deviation of probability judgments

Example

Ψ = [√(.676 / 2),√(.676 / 2),√(.324 / 2),√(.324 / 2)]
γₕ = 4.4045 / √(.5)
γₗ = 0.3306 / √(.5)
σ = .10
model = QOEM(;Ψ, γₕ, γₗ, σ)
predict(model)

References

Trueblood, J. S., & Busemeyer, J. R. (2011). A quantum probability account of order effects in inference. Cognitive science, 35(8), 1518-1552.

Base.rand — Method

rand(model::AbstractQOEM, n::Int; t = 1, r = .05)

Generates simulated data for the following conditions:

Order 1

Pr(disease)

Pr(disease | lab test)

Pr(disease | lab test, medical history)

Order 2

Pr(disease)

Pr(disease | medical history)

Pr(disease | medical history, lab test)

Arguments

model::AbstractQOEM: a model object for a quantum order effect model

n: the number of trials per condition

Keywords

t = 1: time of decision

Example

using QuantumOrderEffectModels
Ψ = @. √([.35,.35,.15,.15])
γₕ = 2
γₗ = .5
σ = .05
n_trials = 100
model = QOEM(;Ψ, γₕ, γₗ, σ)
data = rand(model, n_trials)

Distributions.logpdf — Method

logpdf(model::AbstractQOEM, n::Int, n_d::Vector{Int}; t = 1, r = .05)

Returns the joint log density given data for the following conditions:

Order 1

Pr(disease)

Pr(disease | lab test)

Pr(disease | lab test, medical history)

Order 2

Pr(disease)

Pr(disease | medical history)

Pr(disease | medical history, lab test)

Arguments

model::AbstractQOEM: a model object for a quantum order effect model

n: the number of trials per condition

n_d: the number of defections in each condition

Keywords

`t = 1: time of decision

Example

Ψ = @. √([.35,.35,.15,.15])
γₕ = 2
γₗ = .5
σ = .05
n_trials = 100
model = QOEM(;Ψ, γₕ, γₗ, σ)
data = rand(model, n_trials)
logpdf(model, data)

Distributions.pdf — Method

logpdf(model::AbstractQOEM, n::Int, n_d::Vector{Int}; t = 1, r = .05)

Returns the joint log density given data for the following conditions:

Order 1

Pr(disease)

Pr(disease | lab test)

Pr(disease | lab test, medical history)

Order 2

Pr(disease)

Pr(disease | medical history)

Pr(disease | medical history, lab test)

Arguments

model::AbstractQOEM: a model object for a quantum order effect model

n: the number of trials per condition

n_d: the number of defections in each condition

Keywords

`t = 1: time of decision

Example

using QuantumOrderEffectModels
Ψ = @. √([.35,.35,.15,.15])
γₕ = 2
γₗ = .5
σ = .05
n_trials = 100
model = QOEM(;Ψ, γₕ, γₗ, σ)
data = rand(model, n_trials)
logpdf(model, data)

QuantumOrderEffectModels.predict — Method

predict(model::AbstractQOEM; t = 1)

Returns predicted response probability for the following conditions:

Arguments

model::AbstractQOEM

Keywords

t = 1: time of decision

Output

The output is a vector of predictions corresponding to the following conditions:

Order 1

Pr(disease)

Pr(disease | lab test)

Pr(disease | lab test, medical history)

Order 2

Pr(disease)

Pr(disease | medical history)

Pr(disease | medical history, lab test)

Example

using QuantumOrderEffectModels
Ψ = @. √([.35,.35,.15,.15])
γₕ = 2
γₗ = .5
σ = .05
model = QOEM(;Ψ, γₕ, γₗ, σ)
predict(model)